Fix ARM char portability

+++ src/paimon/common/utils/binary_row_partition_computer_test.cpp
@@ -77,7 +77,7 @@ TEST(BinaryRowPartitionComputerTest, TestToAndFromBinaryRow) {
         ASSERT_OK_AND_ASSIGN(BinaryRow row, computer->ToBinaryRow(partition_map));
         ASSERT_EQ(12, row.GetFieldCount());
         ASSERT_EQ(true, row.GetBoolean(0));
-        ASSERT_EQ(-20, row.GetByte(1));
+        ASSERT_EQ(-20, static_cast<std::int8_t>(row.GetByte(1)));
         ASSERT_EQ(10, row.GetByte(2));
         ASSERT_EQ(1556, row.GetShort(3));
         ASSERT_EQ(-2556, row.GetShort(4));
@@ -121,7 +121,7 @@ TEST(BinaryRowPartitionComputerTest, TestToAndFromBinaryRow) {
         ASSERT_OK_AND_ASSIGN(BinaryRow row, computer->ToBinaryRow(partition_map));
         ASSERT_EQ(12, row.GetFieldCount());
         ASSERT_EQ(true, row.GetBoolean(0));
-        ASSERT_EQ(-20, row.GetByte(1));
+        ASSERT_EQ(-20, static_cast<std::int8_t>(row.GetByte(1)));
         ASSERT_EQ(10, row.GetByte(2));
         ASSERT_EQ(1556, row.GetShort(3));
         ASSERT_EQ(-2556, row.GetShort(4));
@@ -165,7 +165,7 @@ TEST(BinaryRowPartitionComputerTest, TestToAndFromBinaryRow) {
         ASSERT_OK_AND_ASSIGN(BinaryRow row, computer->ToBinaryRow(partition_map));
         ASSERT_EQ(12, row.GetFieldCount());
         ASSERT_EQ(true, row.GetBoolean(0));
-        ASSERT_EQ(-20, row.GetByte(1));
+        ASSERT_EQ(-20, static_cast<std::int8_t>(row.GetByte(1)));
         ASSERT_EQ(10, row.GetByte(2));
         ASSERT_EQ(1556, row.GetShort(3));
         ASSERT_EQ(-2556, row.GetShort(4));
